Logging

See merge request ospr/sose24/group_03!20

+++ src/osmp_library/osmplib.h
@@ -33,7 +33,7 @@
  * Struct für eine Nachricht entsprechend der Definition unseres Shared Memory.
  */
 typedef struct OSMP_message {
-    unsigned short free;                    // Flag: Slot frei (SLOT_FREE) oder nicht (SLOT_TAKEN)?
+    int free;                               // Flag: Slot frei (SLOT_FREE) oder nicht (SLOT_TAKEN)?
     int to;                                 // Rang des empfangenden Prozesses
     int from;                               // Rang des sendenden Prozesses
     int len;                                // Länge der Nachricht in Bytes
